requirements.txt
streamlit
python-dotenv
groq
sentence-transformers
pypdf
faiss-cpu
numpy